some layout changes across the codebase
