+++ src/app/(dashboard)/layout.tsx
@@ -7,14 +7,10 @@ export default function DashboardLayout({
 }>) {
   return (
     <>
-      <header className="flex h-16 shrink-0 items-center gap-2 transition-[width,height] ease-linear group-has-[[data-collapsible=icon]]/sidebar-wrapper:h-12">
-        <Navbar />
-      </header>
       {/* MAIN PANEL - dashboard panels*/}
       <main className="h-full w-full flex bg-[#f7f8fa] p-3">
         {children}
       </main>
     </>
-
   )
 }
